Separate system and encrypted config methods
