+++ server/src/controllers/config/encrypted.ts
@@ -1,0 +1,153 @@
+const path = require("path");
+const { logEvent } = require("../../utils/logger");
+const { PATHS, isValidAppName } = require("../../config");
+const { encryptValue, decryptValue } = require("../../utils/encryption");
+import { Request, Response } from "express";
+
+// App-related interfaces start here
+interface UpdateAppConfigRequestParams {
+  appName: string;
+}
+
+interface UpdateAppConfigRequestBody {
+  config: Record<string, any>;
+}
+
+interface UpdateAppConfigResponse {
+  appName: string;
+  config: Record<string, any>;
+  message: string;
+}
+
+interface AppConfigErrorResponse {
+  error: string;
+  details?: string;
+}
+
+const updateAppConfig = async (
+  req: Request<UpdateAppConfigRequestParams, {}, UpdateAppConfigRequestBody>,
+  res: Response<UpdateAppConfigResponse | AppConfigErrorResponse>
+): Promise<void> => {
+  // Implementation will go here
+};
+
+/**
+ * Retrieves all encrypted configuration values for an application.
+ *
+ * @param req - The request object containing the application name in the params.
+ * @param res - The response object used to send back the desired HTTP response.
+ * @returns {Promise<void>} - A promise that resolves when the encrypted configuration values are retrieved.
+ */
+interface GetAppEncryptedConfigRequestParams {
+  appName: string;
+}
+
+interface AppEncryptedConfigResponse {
+  appName: string;
+  encrypted: Record<string, string>;
+}
+
+const getAppEncryptedConfig = async (
+  req: Request<GetAppEncryptedConfigRequestParams>,
+  res: Response<AppEncryptedConfigResponse | AppConfigErrorResponse>
+): Promise<void> => {
+  // Implementation will go here
+};
+
+/**
+ * Creates or updates multiple encrypted configuration values for an application.
+ *
+ * @param req - The request object containing the application name in params and encrypted values in body.
+ * @param res - The response object used to send back the desired HTTP response.
+ * @returns {Promise<void>} - A promise that resolves when the encrypted configuration values are updated.
+ */
+interface CreateAppEncryptedConfigRequestParams {
+  appName: string;
+}
+
+interface CreateAppEncryptedConfigRequestBody {
+  encrypted: Record<string, string>;
+}
+
+interface CreateAppEncryptedConfigResponse {
+  appName: string;
+  keys: string[];
+  message: string;
+}
+
+const createAppEncryptedConfig = async (
+  req: Request<
+    CreateAppEncryptedConfigRequestParams,
+    {},
+    CreateAppEncryptedConfigRequestBody
+  >,
+  res: Response<CreateAppEncryptedConfigResponse | AppConfigErrorResponse>
+): Promise<void> => {
+  // Implementation will go here
+};
+
+/**
+ * Creates or updates a specific encrypted configuration value for an application.
+ *
+ * @param req - The request object containing the application name and key in params, and value in body.
+ * @param res - The response object used to send back the desired HTTP response.
+ * @returns {Promise<void>} - A promise that resolves when the encrypted configuration value is updated.
+ */
+interface UpdateAppEncryptedValueRequestParams {
+  appName: string;
+  key: string;
+}
+
+interface UpdateAppEncryptedValueRequestBody {
+  value: string;
+}
+
+interface UpdateAppEncryptedValueResponse {
+  appName: string;
+  key: string;
+  message: string;
+}
+
+const updateAppEncryptedValue = async (
+  req: Request<
+    UpdateAppEncryptedValueRequestParams,
+    {},
+    UpdateAppEncryptedValueRequestBody
+  >,
+  res: Response<UpdateAppEncryptedValueResponse | AppConfigErrorResponse>
+): Promise<void> => {
+  // Implementation will go here
+};
+
+/**
+ * Deletes a specific encrypted configuration value for an application.
+ *
+ * @param req - The request object containing the application name and key in params.
+ * @param res - The response object used to send back the desired HTTP response.
+ * @returns {Promise<void>} - A promise that resolves when the encrypted configuration value is deleted.
+ */
+interface DeleteAppEncryptedValueRequestParams {
+  appName: string;
+  key: string;
+}
+
+interface DeleteAppEncryptedValueResponse {
+  appName: string;
+  key: string;
+  message: string;
+}
+
+const deleteAppEncryptedValue = async (
+  req: Request<DeleteAppEncryptedValueRequestParams>,
+  res: Response<DeleteAppEncryptedValueResponse | AppConfigErrorResponse>
+): Promise<void> => {
+  // Implementation will go here
+};
+
+// Export using CommonJS syntax
+module.exports = {
+  getAppEncryptedConfig,
+  createAppEncryptedConfig,
+  updateAppEncryptedValue,
+  deleteAppEncryptedValue,
+};
