+++ server/src/controllers/config/index.ts
@@ -1,12 +1,9 @@
 // This file acts as an aggregator for the config controller functions.
-
-const configController = require("./config");
 const system = require("./system");
 const app = require("./app");
 const encrypted = require("./encrypted");
 
 module.exports = {
-  ...configController,
   ...system,
   ...app,
   ...encrypted,
